+++ lib/eidmsdk_platform_interface.dart
@@ -45,8 +45,7 @@ abstract class EidmsdkPlatform extends PlatformInterface {
   }
 
   Future<CertificatesInfo?> getCertificates({
-    // TODO Change type to single EIDCertificateIndex because Android supports only single element
-    required List<EIDCertificateIndex> types,
+    required EIDCertificateIndex type,
     String? language,
   }) {
     throw UnimplementedError('getCertificates() has not been implemented.');